+++ CHANGELOG.md
@@ -10,13 +10,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
-- `telemetry.distro.name` resource attribute with the value `splunk-otel-go`.
-- `telemetry.distro.version` resource attribute with the value of the current
+- Add `telemetry.distro.name` resource attribute with the value `splunk-otel-go`.
+- Add `telemetry.distro.version` resource attribute with the value of the current
   release version.
 
 ### Deprecated
 
-- `splunk.distro.version` resource attribute
+- `splunk.distro.version` resource attribute is deprecated and may be removed
+  in a future release.
 
 ## [1.25.0] - 2025-03-20