+++ CHANGELOG.md
@@ -8,6 +8,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [1.26.0] - 2025-05-21
+
+This release upgrades [OpenTelemetry Go to v1.36.0/v0.58.0][otel-v1.35.0] and [v0.12.1][otel-log-v0.12.1]
+and [OpenTelemetry Go Contrib to v1.36.0/v0.61.0/v0.30.0/v0.16.0/v0.11.0/v0.9.0/v0.8.0][contrib-v1.36.0].
+
 ### Added
 
 - Add `telemetry.distro.name` resource attribute with the value `splunk-otel-go`.
@@ -661,7 +666,8 @@ an impedance mismatch with this duplicate batching.
 - Add [`splunkhttp`](./instrumentation/net/http/splunkhttp) module providing
   additional Splunk specific instrumentation for `net/http`.
 
-[Unreleased]: https://github.com/signalfx/splunk-otel-go/compare/v1.25.0...HEAD
+[Unreleased]: https://github.com/signalfx/splunk-otel-go/compare/v1.26.0...HEAD
+[1.26.0]: https://github.com/signalfx/splunk-otel-go/releases/tag/v1.26.0
 [1.25.0]: https://github.com/signalfx/splunk-otel-go/releases/tag/v1.25.0
 [1.24.0]: https://github.com/signalfx/splunk-otel-go/releases/tag/v1.24.0
 [1.23.0]: https://github.com/signalfx/splunk-otel-go/releases/tag/v1.23.0
@@ -699,6 +705,8 @@ an impedance mismatch with this duplicate batching.
 [0.2.0]: https://github.com/signalfx/splunk-otel-go/releases/tag/v0.2.0
 [0.1.0]: https://github.com/signalfx/splunk-otel-go/releases/tag/v0.1.0
 
+[otel-log-v0.12.1]: https://github.com/open-telemetry/opentelemetry-go/releases/tag/log/v0.12.1
+[otel-v1.36.0]: https://github.com/open-telemetry/opentelemetry-go/releases/tag/v1.36.0
 [otel-v1.35.0]: https://github.com/open-telemetry/opentelemetry-go/releases/tag/v1.35.0
 [otel-v1.34.0]: https://github.com/open-telemetry/opentelemetry-go/releases/tag/v1.34.0
 [otel-v1.33.0]: https://github.com/open-telemetry/opentelemetry-go/releases/tag/v1.33.0
@@ -734,6 +742,7 @@ an impedance mismatch with this duplicate batching.
 [otel-v0.20.0]: https://github.com/open-telemetry/opentelemetry-go/releases/tag/v0.20.0
 [otel-v0.19.0]: https://github.com/open-telemetry/opentelemetry-go/releases/tag/v0.19.0
 
+[contrib-v1.36.0]: https://github.com/open-telemetry/opentelemetry-go-contrib/releases/tag/v1.36.0
 [contrib-v1.35.0]: https://github.com/open-telemetry/opentelemetry-go-contrib/releases/tag/v1.35.0
 [contrib-v1.34.0]: https://github.com/open-telemetry/opentelemetry-go-contrib/releases/tag/v1.34.0
 [contrib-v1.33.0]: https://github.com/open-telemetry/opentelemetry-go-contrib/releases/tag/v1.33.0